+++ src/components/Button/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { StyledButton } from './styles';
+
+export default function Button({ onClick, text, ...props }) {
+  return (
+    <StyledButton {...props} onClick={onClick}>
+      {text}
+    </StyledButton>
+  );
+}
