+++ src/components/Nav/styles.js
@@ -1,21 +1,12 @@
-import React from "react";
-import { NavLink } from "react-router-dom";
-import styled from "@emotion/styled";
+import { NavLink } from 'react-router-dom';
+import styled from '@emotion/styled';
 
-export default function Nav() {
-  return (
-    <NavContainer>
-      <Menu to='/'>지도 → 폴리곤</Menu>
-      <Menu to='/draw'>폴리곤 → 지도</Menu>
-    </NavContainer>
-  );
-}
-
-const NavContainer = styled.nav`
-  font-family: "Fredoka", sans-serif;
+export const NavContainer = styled.nav`
+  font-family: 'Fredoka', sans-serif;
   margin-bottom: 20px;
 `;
-const Menu = styled(NavLink)`
+
+export const Menu = styled(NavLink)`
   font-size: 18px;
   color: royalblue;
   text-decoration: none;
@@ -27,7 +18,7 @@ const Menu = styled(NavLink)`
     font-weight: 600;
     position: relative;
     &::after {
-      content: "";
+      content: '';
       width: 100%;
       height: 6px;
       background: #e0e0ff;
