+++ src/components/Option/index.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { StyledButton } from './styles';
+
+export default function Option({ onClick, text, ...props }) {
+  return <StyledButton {...props} onClick={onClick} >{text}</StyledButton>;
+}
