+++ src/components/Option/styles.js
@@ -1,0 +1,16 @@
+import styled from '@emotion/styled';
+
+export const StyledButton = styled.button`
+  border: 0;
+  border-radius: 20px;
+  padding: 5px 15px;
+  color: royalblue;
+  background-color: ${(props) => (props.target ? '#f6f6f6' : '#fff')};
+  font-weight: ${(props) => props.target && '700'};
+  font-family: 'Fredoka', sans-serif;
+  cursor: pointer;
+  z-index: 0;
+  &:hover {
+    font-weight: 700;
+  }
+`;
