+++ src/container/input/index.jsx
@@ -1,10 +1,12 @@
 import React, { useState } from 'react';
+//TODO: module.css에서 emotion으로 완전히 변경
 import styles from './DrawContainer.module.css';
-import DrawNaverMap from './DrawNaverMap.js';
-import InputPolygon from './InputPolygon.js';
-import Nav from '../Nav';
+import DrawNaverMap from './NaverMap.jsx';
+import InputPolygon from './InputPolygon.jsx';
+import Nav from '../../components/Nav';
 
 function DrawContainer() {
+  // TODO: state들 상태관리 라이브러리로 분리
   const [polygon, setPolygon] = useState([]);
   const [option, setOption] = useState('WKT');
   // eslint-disable-next-line no-useless-escape
