+++ src/store/click/index.js
@@ -1,6 +1,6 @@
 import create from 'zustand';
 
-const useStore = create((set) => ({
+export const useStore = create((set) => ({
   polygon: '',
   marker: [],
   clickPolygon: ({ coord: { x, y } }) =>
