+++ src/store/input/index.js
@@ -1,0 +1,7 @@
+import create from 'zustand';
+
+export const useStore = create((set) => ({
+  //TODO: 채워넣기
+}));
+
+export default useStore;
